+++ BuildPropertiesPlugin/plugin/src/main/groovy/com/novoda/buildproperties/BuildPropertiesPlugin.groovy
@@ -22,6 +22,7 @@ class BuildPropertiesPlugin implements Plugin<Project> {
                 return new BuildProperties(name, project)
             }
         })
+        container.create('env').entries(new EnvironmentPropertiesEntries(project))
         project.extensions.add('buildProperties', container)
 
         def android = project.extensions.findByName("android")
@@ -65,28 +66,11 @@ class BuildPropertiesPlugin implements Plugin<Project> {
         target.ext.buildConfigString = { String name, String value ->
             buildConfigField('String', name, { "\"$value\"" })
         }
-        target.ext.buildConfigProperty = { String name = null, BuildProperties.Entry entry ->
+        target.ext.buildConfigProperty = { String name, Entry entry ->
             project.afterEvaluate {
-                target.buildConfigField 'String', name ?: formatBuildConfigField(entry.key), "\"${entry.string}\""
+                target.buildConfigField 'String', name, "\"${entry.string}\""
             }
         }
-        target.ext.buildConfigProperties = { BuildProperties buildProperties ->
-            buildProperties.keys.each { String key ->
-                target.ext.buildConfigProperty buildProperties[key]
-            }
-        }
-    }
-
-    private String formatBuildConfigField(String name) {
-        return splitTokens(name)
-                .toUpperCase()
-    }
-
-    private String splitTokens(String name) {
-        return name
-                .replace('.', '_')
-                .split("(?<!(^|[A-Z]))(?=[A-Z])|(?<!^)(?=[A-Z][a-z])")
-                .join('_')
     }
 
     private void addResValueSupportTo(target, Project project) {
@@ -102,31 +86,19 @@ class BuildPropertiesPlugin implements Plugin<Project> {
         target.ext.resValueString = { String name, String value ->
             resValue('string', name, { "\"$value\"" })
         }
-        target.ext.resValueProperty = { String name = null, BuildProperties.Entry entry ->
+        target.ext.resValueProperty = { String name, Entry entry ->
             project.afterEvaluate {
-                target.resValue 'string', name ?: formatResValueName(entry.key), "\"${entry.string}\""
+                target.resValue 'string', name, "\"${entry.string}\""
             }
         }
-        target.ext.resValueProperties = { BuildProperties buildProperties ->
-            buildProperties.keys.each { String key ->
-                target.ext.resValueProperty buildProperties[key]
-            }
-        }
-    }
-
-    private String formatResValueName(String name) {
-        return splitTokens(name)
-                .toLowerCase()
     }
 
     private void addSigningConfigSupportTo(target, Project project) {
         target.ext.signingConfigProperties = { BuildProperties buildProperties ->
-            project.afterEvaluate {
-                target.storeFile new File(buildProperties.parentFile, buildProperties['storeFile'].string)
-                target.storePassword buildProperties['storePassword'].string
-                target.keyAlias buildProperties['keyAlias'].string
-                target.keyPassword buildProperties['keyPassword'].string
-            }
+            target.storeFile new File(buildProperties.parentFile, buildProperties['storeFile'].string)
+            target.storePassword buildProperties['storePassword'].string
+            target.keyAlias buildProperties['keyAlias'].string
+            target.keyPassword buildProperties['keyPassword'].string
         }
     }