+++ slacker/public/widgets/coverage.js
@@ -3,7 +3,7 @@ if (widgets.coverage) throw "Coverage already exists";
 widgets.coverage = function() {
   return {
     present: function(data, element) {
-      element.innerHTML = presentCoverage(data.payload[Math.floor((Math.random() * data.payload.length))]);
+      element.innerHTML = presentCoverage(data.payload);
     }
   };
 }
@@ -12,7 +12,7 @@ function presentCoverage(data) {
   return '<div class="content-container active-image">' +
   '<div class="active-container">' +
     '<div class="active-filler"></div>' +
-    '<span class="active-channel-header">' + '&#128202;' + '</span>' +
+    '<span class="active-channel-header">' + '&#x1f4ca;' + '</span>' +
     '<br>' +
     '<div>' +
       '<span class="active-channel-copy"><b class="active-channel-name">#' + data.project + '</b><i> has ' + data.coverage + '% coverage</i></span>' +