+++ slacker/server/coverage.js
@@ -4,77 +4,90 @@ module.exports = {
 }
 
 const sonarAuth = process.env.SONAR_AUTH;
-var httpClient = require('request');
+const PROJECTS_REQUEST = {
+  url: 'https://sonar.novoda.com/api/projects/index',
+  auth: {
+    user: sonarAuth,
+    pass: '',
+  }
+};
+const httpClient = require('request-promise-native');
+
+var currentIndex = 0;
 
 function coverage() {
-  return new Promise(getProjects).then(toAllCoverage).then(toRuleResult);
+  return httpClient(PROJECTS_REQUEST)
+    .then(parseProjects)
+    .then(toAllCoverage)
+    .then(toRuleResult);
 }
 
-function getProjects(resolve, reject) {
-  var projectsRequest = {
-    url: 'https://sonar.novoda.com/api/projects/index',
+function parseProjects(body) {
+  var jsonBody = JSON.parse(body);
+  var results = jsonBody.map(each => {
+    return {
+      name: each.nm,
+      key: each.k
+    }
+  });
+  return Promise.resolve(results);
+}
+
+function toAllCoverage(data) {
+  var allCoverage = data.map(each => {
+    var coverageRequest = createCoverageRequest(each.key);
+    return httpClient(coverageRequest)
+      .then(parseCoverage(each.name));
+  });
+  return Promise.all(allCoverage);
+}
+
+function createCoverageRequest(key) {
+  return {
+    url:  'https://sonar.novoda.com/api/measures/component',
+    qs: {
+      componentKey: key,
+      metricKeys: 'coverage'
+    },
     auth: {
       user: sonarAuth,
       pass: '',
     }
   };
+}
 
-  httpClient.get(projectsRequest, function(error, response, body) {
+function parseCoverage(projectName) {
+  return function(body) {
     var jsonBody = JSON.parse(body);
-    var results = jsonBody.map(each => {
-      return {
-        name: each.nm,
-        key: each.k
-      }
-    });
-    resolve(results);
-  });
-}
-
-function toAllCoverage(data) {
-  var allCoverage = data.map(each => {
-    return toCoverage(each);
-  });
-  return Promise.all(allCoverage);
-}
-
-function toCoverage(data) {
-  var result = function(resolve, reject) {
-    var coverageRequest = {
-      url:  'https://sonar.novoda.com/api/measures/component',
-      qs: {
-        componentKey: data.key,
-        metricKeys: 'coverage'
-      },
-      auth: {
-        user: sonarAuth,
-        pass: '',
-      }
-    };
-    httpClient.get(coverageRequest, function(error, response, body) {
-      var jsonBody = JSON.parse(body);
-      var measures = jsonBody.component.measures;
-      if (measures && measures.length > 0) {
-        resolve({
-          project: data.name,
-          coverage: measures[0].value
-        });
-      } else {
-        resolve();
-      }
-    });
-  };
-  return new Promise(result);
+    var measures = jsonBody.component.measures;
+    if (measures && measures.length > 0) {
+      return Promise.resolve({
+        project: projectName,
+        coverage: measures[0].value
+      });
+    } else {
+      return Promise.resolve({});
+    }
+  }
 }
 
 function toRuleResult(data) {
   var filtered = data.filter(each => {
-    return each;
+    return each.project;
   });
-  return new Promise(function(resolve, reject) {
-    resolve({
-      widgetKey: 'coverage',
-      payload: filtered
-    });
+
+  incrementIndex(filtered.length);
+
+  return Promise.resolve({
+    widgetKey: 'coverage',
+    payload: filtered[currentIndex]
   });
+}
+
+function incrementIndex(dataLength) {
+  if (currentIndex >= dataLength) {
+    currentIndex = 0;
+  } else {
+    currentIndex++;
+  }
 }