+++ slacker/server/dashboard.js
@@ -1,6 +1,6 @@
 var Slacker = require('./slacker/slacker.js');
 
-const DASHBOARD_INTERVAL = 1000 * 15;
+const DASHBOARD_INTERVAL = 1000 * 30;
 
 var Dashboard = function(token) {
   this.slacker = new Slacker(token);
@@ -47,10 +47,6 @@ function incrementIndex(self) {
 
 function getTimeoutInterval(self) {
   return DASHBOARD_INTERVAL * self.widgets[self.index].rank
-}
-
-Dashboard.prototype.forceUpdate = function(callback) {
-  //update(this, callback);
 }
 
 module.exports = Dashboard;