+++ BuildPropertiesPlugin/CHANGELOG.md
@@ -1,13 +1,32 @@
 Change Log
 ==========
 
+Version 1.2.1 *(18/09/2016)*
+----------------------------
+
+- Fixed bug with `signingConfigProperties` not filling signing configuration
+ as intended
+
+Version 1.2.0 *(03/09/2016)*
+----------------------------
+
+- Added lazy-loading for properties files
+
+Version 1.1.0 *(25/08/2016)*
+----------------------------
+
+- Removed `buildConfigProperties` / `resValueProperties`
+- Added fallback support to `Entry` via `or()` operator
+- Added environment variables support via built-in `buildProperties.env` set
+
+
 Version 1.0.1 *(16/07/2016)*
 ----------------------------
 
-Added possibility to specify error message when properties file not found.
+- Added possibility to specify error message when properties file not found.
 
 
 Version 1.0.0 *(16/07/2016)*
 ----------------------------
 
-Initial release.
+- Initial release.